refactor: rename file rigid_transforms into solvers
